Refactoring Authors code for better error handling.

+++ LibraryManagementSystem.Web/Controllers/AuthorController.cs
@@ -1,5 +1,4 @@
-﻿using LibraryManagementSystem.Services.Data;
-using LibraryManagementSystem.Services.Data.Interfaces;
+﻿using LibraryManagementSystem.Services.Data.Interfaces;
 using LibraryManagementSystem.Services.Data.Models.Author;
 using LibraryManagementSystem.Web.ViewModels.Author;
 using Microsoft.AspNetCore.Authorization;
@@ -18,6 +17,7 @@ namespace LibraryManagementSystem.Web.Controllers
             this.authorService = authorService;
         }
 
+        // ready
         [HttpGet]
         [Authorize(Roles = AdminRole)]
         public IActionResult Add()
@@ -27,19 +27,30 @@ namespace LibraryManagementSystem.Web.Controllers
             return View(author);
         }
 
+        // ready
         [HttpPost]
         [Authorize(Roles = AdminRole)]
         public async Task<IActionResult> Add(AuthorFormModel model)
         {
-            if (!ModelState.IsValid)
+            if (!this.ModelState.IsValid)
             {
-                return View(model);
+                return this.View(model);
             }
 
             try
             {
+                bool doesAuthorExist = await this.authorService.AuthorExistByNameAndNationalityAsync(model.FirstName, model.LastName, model.Nationality);
+
+                if (doesAuthorExist)
+                {
+                    TempData[ErrorMessage] = "Author with the same Name and Nationality already exists!";
+
+                    return this.RedirectToAction("All", "Author");
+                }
+
                 await authorService.AddAuthorAsync(model);
-                TempData[SuccessMessage] = "Succestully added author";
+
+                TempData[SuccessMessage] = "Successfully added author.";
             }
             catch
             {
@@ -59,6 +70,7 @@ namespace LibraryManagementSystem.Web.Controllers
         //    return View(viewModel);
         //}
 
+        // ready
         [HttpGet]
         [AllowAnonymous]
         public async Task<IActionResult> All([FromQuery] AllAuthorsQueryModel queryModel)
@@ -71,20 +83,23 @@ namespace LibraryManagementSystem.Web.Controllers
             return this.View(queryModel);
         }
 
+        // ready
         [HttpGet]
         [Authorize(Roles = AdminRole)]
         public async Task<IActionResult> Edit(string id)
         {
             try
             {
-                var author = await authorService.GetAuthorForEditByIdAsync(id);
+                bool authorExists = await this.authorService.AuthorExistByIdAsync(id);
 
-                if (author == null)
+                if (!authorExists)
                 {
-                    this.TempData[ErrorMessage] = "Such author does not exists!";
+                    TempData[ErrorMessage] = "Such author does not exists!";
 
                     return RedirectToAction("All", "Author");
                 }
+
+                var author = await authorService.GetAuthorForEditByIdAsync(id);
 
                 return View(author);
             }
@@ -94,28 +109,30 @@ namespace LibraryManagementSystem.Web.Controllers
             }
         }
 
+        // ready
         [HttpPost]
         [Authorize(Roles = AdminRole)]
         public async Task<IActionResult> Edit(string id, AuthorFormModel model)
         {
             if (!ModelState.IsValid)
             {
-                this.TempData[ErrorMessage] = "Such author does not exists!";
-
                 return View(model);
+            }
+
+            bool authorExists = await this.authorService.AuthorExistByIdAsync(id);
+
+            if (!authorExists)
+            {
+                TempData[ErrorMessage] = "Such author does not exists!";
+
+                return RedirectToAction("All", "Author");
             }
 
             try
             {
-                var author = await authorService.GetAuthorForEditByIdAsync(id);
-
-                if (author == null)
-                {
-                    TempData[ErrorMessage] = "There is no author with such id!";
-                }
-
                 await authorService.EditAuthorAsync(id, model);
-                TempData[SuccessMessage] = "Succesfully edited author";
+
+                TempData[SuccessMessage] = "Succesfully edited author.";
             }
             catch
             {
@@ -125,34 +142,24 @@ namespace LibraryManagementSystem.Web.Controllers
             return this.RedirectToAction("All", "Author");
         }
 
-        [HttpGet]
-        [Authorize(Roles = "Administrator, User")]
-        public async Task<IActionResult> Details(string id)
-        {
-            try
-            {
-                AuthorDetailsViewModel? author = await authorService.GetAuthorDetailsForUserAsync(id);
-
-                if (author == null)
-                {
-                    return RedirectToAction("All", "Author");
-                }
-
-                return View(author);
-            }
-            catch
-            {
-                return GeneralError();
-            }
-        }
-
+        // ready
         [HttpGet]
         [Authorize(Roles = AdminRole)]
         public async Task<IActionResult> Delete(string id)
         {
+            bool authorExists = await this.authorService.AuthorExistByIdAsync(id);
+
+            if (!authorExists)
+            {
+                TempData[ErrorMessage] = "Such author does not exists!";
+
+                return RedirectToAction("All", "Author");
+            }
+
             try
             {
                 await authorService.DeleteAuthorAsync(id);
+                TempData[SuccessMessage] = "Succesfully deleted author.";
             }
             catch
             {
@@ -160,6 +167,31 @@ namespace LibraryManagementSystem.Web.Controllers
             }
 
             return this.RedirectToAction("All", "Author");
+        }
+    
+        [HttpGet]
+        [Authorize(Roles = "Administrator, User")]
+        public async Task<IActionResult> Details(string id)
+        {
+            bool authorExists = await this.authorService.AuthorExistByIdAsync(id);
+
+            if (!authorExists)
+            {
+                TempData[ErrorMessage] = "Such author does not exists!";
+
+                return RedirectToAction("All", "Author");
+            }
+
+            try
+            {
+                AuthorDetailsViewModel? author = await authorService.GetAuthorDetailsForUserAsync(id);
+
+                return View(author);
+            }
+            catch
+            {
+                return GeneralError();
+            }
         }
 
         private IActionResult GeneralError()
